Restructure and make more compact
